+++ apps/views/partials/file-upload-form/input-file.html
@@ -6,9 +6,15 @@
     <span id="validation-error" class="govuk-visually-hidden">{{#t}}journey.error{{/t}}:</span> {{#t}}validation.file-upload.maxFileSize{{/t}}
   </p>
   <p id="file-upload-error-fileType" class="govuk-error-message govuk-!-display-none">
-    <span id="validation-error" class="govuk-visually-hidden">{{#t}}journey.error{{/t}}:</span> {{#t}}validation.file-upload.fileType{{/t}}
+    <span id="validation-error" class="govuk-visually-hidden">{{#t}}journey.error{{/t}}:</span> 
+    {{#documentCategory.customFileType}}
+      {{#t}}validation.file-upload.{{documentCategory.name}}.fileType{{/t}}
+    {{/documentCategory.customFileType}}
+    {{^documentCategory.customFileType}}
+      {{#t}}validation.file-upload.fileType{{/t}}
+    {{/documentCategory.customFileType}}
   </p>
-  <input class="govuk-file-upload" id="file-upload" name="file-upload" type="file" document-category="{{options.documentCategory}}">
+  <input class="govuk-file-upload" id="file-upload" name="file-upload" type="file" document-category="{{documentCategory.name}}">
   <div id="upload-page-loading-spinner" class="spinner-container">
     <div class="spinner-loader"></div>
     <span class="spinner-message">{{#t}}journey.uploading{{/t}}</span>
