+++ assets/js/index.js
@@ -3,7 +3,7 @@
 
 require('hof/frontend/themes/gov-uk/client-js');
 const accessibleAutocomplete = require('accessible-autocomplete');
-const config = require('../../config');
+const { initFileUpload } = require('./file-upload');
 
 document.querySelectorAll('.typeahead').forEach(function applyTypeahead(element) {
   accessibleAutocomplete.enhanceSelectElement({
@@ -15,10 +15,6 @@ document.querySelectorAll('.typeahead').forEach(function applyTypeahead(element)
 document.addEventListener('DOMContentLoaded', () => {
   const loaderContainer = document.querySelector('#loader-container');
   const reportSubmitButton = document.querySelector('#report-submit');
-  const fileUpload = document.getElementById('file-upload');
-  const uploadPageLoaderContainer = document.getElementById('upload-page-loading-spinner');
-  const continueWithoutUpload = document.getElementsByName('continueWithoutUpload');
-  const removeLinks = document.querySelectorAll('#uploaded-documents > div > div > a');
 
   if (loaderContainer) {
     document.querySelector('#report-submit .govuk-button').addEventListener('click', () => {
@@ -27,70 +23,6 @@ document.addEventListener('DOMContentLoaded', () => {
     });
   }
 
-  const fileUploadStatusHandler = (status, errorType) => {
-    const fileUploadComponent = document.getElementById('hofFileUpload');
-    const fileUploadErrorMsg = fileUploadComponent.querySelectorAll('.govuk-error-message');
-    switch (status) {
-      case 'ready':
-        if (fileUploadComponent) {
-          fileUploadComponent.classList.remove('govuk-form-group--error');
-        }
-        if (fileUploadErrorMsg) {
-          fileUploadErrorMsg.forEach(element => {
-            element.classList.add('govuk-!-display-none');
-          });
-        }
-        break;
-      case 'error':
-        if (fileUploadComponent) {
-          fileUploadComponent.classList.add('govuk-form-group--error');
-          document.getElementById(`file-upload-error-${errorType}`).classList.remove('govuk-!-display-none');
-        }
-        break;
-      case 'uploading':
-        uploadPageLoaderContainer.style.display = 'flex';
-        fileUpload.disabled = true;
-        fileUpload.ariaDisabled = true;
-        continueWithoutUpload.forEach(a => {
-          a.disabled = true;
-          a.ariaDisabled = true;
-        });
-        removeLinks.forEach(a => {
-          a.classList.add('disabled-link');
-        });
-        break;
-      default:
-        break;
-    }
-  };
-
-  if (fileUpload) {
-    fileUpload.addEventListener('change', () => {
-      fileUploadStatusHandler('ready');
-      const fileInfo = fileUpload.files && fileUpload.files.length > 0 ? fileUpload.files[0] : null;
-
-      // Retrieve the document-category attribute from the file input field
-      const documentCategory = fileUpload.getAttribute('document-category');
-
-      if (fileInfo) {
-        if (fileInfo.size > config.upload.maxFileSizeInBytes) {
-          fileUploadStatusHandler('error', 'maxFileSize');
-          return;
-        }
-
-        // Get allowedMimeTypes for the documentCategory or fallback to global allowedMimeTypes
-        const documentCategoryConfig = config.upload.documentCategories[documentCategory];
-        const allowedMimeTypes =
-          (documentCategoryConfig && documentCategoryConfig.allowedMimeTypes) || config.upload.allowedMimeTypes;
-
-        if (!allowedMimeTypes.includes(fileInfo.type)) {
-          fileUploadStatusHandler('error', 'fileType');
-          return;
-        }
-
-        document.querySelector('[name=file-upload-form]').submit();
-        fileUploadStatusHandler('uploading');
-      }
-    });
-  }
+  // Initialise file upload logic
+  initFileUpload();
 });
